random - adds kepyboarder test removes movement test from player checkShotFired varied slightly to start with count of 0 instead of -1 objective and instructions added to html board height slightly reduced
