+++ index.html
@@ -6,7 +6,10 @@
 </head>
 <body>
   <h1 id="title">Meteor Shower!</h1>
-  <canvas id="screen" width="450" height="450" style="border:3px solid #777;"></canvas>
+  <h3 id="objective">Meteors are crashing to Earth! Use your Cannon to take them out before they hit the ground!</h3>
+  <h4 id="controls">Controls: MOVE with left and right arrows, FIRE by holding spacebar</h4>
+  <h4 id="tip">Tip: Smaller meteors are worth more points. As you accumulate points you will move through levels and the game will get harder!</h4>
+  <canvas id="screen" width="450" height="400" style="border:3px solid #777;"></canvas>
   <h2 id="score"></h2>
   <h2 id="meteors-destroyed"></h2>
