+++ lib/player.js
@@ -6,7 +6,7 @@ function Player(board) {
   this.size = { width: 16, height: 16 };
   this.velocity = { left: -7, right: 7 };
   this.keyboarder = new Keyboarder();
-  this.rateOfFireCounter = -1;
+  this.rateOfFireCounter = 0;
   this.rateOfFireValue = 15;
   this.active = true;
   this.board.players.push(this);
@@ -51,10 +51,10 @@ Player.prototype.checkRightEdge = function(rightEdge) {
 
 Player.prototype.checkShotFired = function() {
   if (this.keyboarder.isDown(this.keyboarder.KEYS.SPACE)) {
-    this.rateOfFireCounter++;
     if (this.rateOfFireCounter % this.rateOfFireValue === 0) {
       var bullet = this.board.addBullet();
     }
+    this.rateOfFireCounter++;
   }
 };
