+++ test/index.js
@@ -4,3 +4,4 @@ require('./meteor-test');
 require('./bullet-test');
 require('./score-test');
 require('./collision-test');
+require('./keyboarder-test');
